requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
astropy==5.1
    # via -r requirements.in
autopep8==1.6.0
    # via -r requirements.in
click==8.1.3
    # via pip-tools
cycler==0.11.0
    # via matplotlib
flake8==4.0.1
    # via -r requirements.in
fonttools==4.33.3
    # via matplotlib
h5py==3.7.0
    # via yt-astro-analysis
kiwisolver==1.4.2
    # via matplotlib
matplotlib==3.5.2
    # via -r requirements.in
mccabe==0.6.1
    # via flake8
more-itertools==8.13.0
    # via yt-astro-analysis
mpmath==1.2.1
    # via sympy
numpy==1.22.4
    # via
    #   -r requirements.in
    #   astropy
    #   colorspacious
    #   h5py
    #   matplotlib
    #   pyerfa
    #   scipy
    #   unyt
packaging==21.3
    # via
    #   astropy
    #   matplotlib
pep517==0.12.0
    # via pip-tools
pillow==9.1.1
    # via matplotlib
pip-tools==6.6.2
    # via -r requirements.in
pycodestyle==2.8.0
    # via
    #   autopep8
    #   flake8
pyerfa==2.0.0.1
    # via astropy
pyflakes==2.4.0
    # via flake8
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via matplotlib
pyyaml==6.0
    # via astropy
scipy==1.8.1
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sympy==1.10.1
    # via unyt
toml==0.10.2
    # via autopep8
tomli==2.0.1
    # via pep517
unyt==2.8.0
    # via -r requirements.in
wheel==0.37.1
    # via pip-tools
yt @ git+https://github.com/brittonsmith/yt@gadget4
    # via
    #   -r requirements.in
    #   yt-astro-analysis
    #   yt-legacy
yt-astro-analysis==1.1.1
    # via -r requirements.in
yt-legacy @ git+https://github.com/brittonsmith/yt_legacy@master
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
